requirements.txt
aiohttp==3.8.4
tenacity==8.2.2
xmltodict==0.13.0
tabulate==0.9.0
python-dotenv==1.0.0
pyyaml==6.0
dnspython==2.3.0